pyproject.toml
[project]
name = "pix-notification"
version = "0.1.0"
description = ""
authors = [
    {name = "guilherme",email = "guilherme-lopes58@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "pix_notification", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"


[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
